requirements.txt
pandas
numpy
scikit-learn

nltk
rank-bm25
sentence-transformers
faiss-cpu
google-generativeai

fastapi
uvicorn
pydantic

python-multipart
starlette

streamlit
requests
python-dotenv

selenium
